Move temple transition to stage config
